improve typings
