+++ packages/zent/typings/libs/Badge.d.ts
@@ -1,0 +1,14 @@
+/// <reference types="react" />
+
+declare module 'zent/lib/badge' {
+  interface IBadgeProps {
+    count?: number
+    maxCount?: number
+    dot?: boolean
+    showZero?: boolean
+    className?: string
+    prefix?: string
+  }
+
+  export class Badge extends React.Component<IBadgeProps, any> {}
+}
