+++ packages/zent/typings/libs/BlockHeader.d.ts
@@ -1,0 +1,14 @@
+/// <reference types="react" />
+
+declare module 'zent/lib/block-header' {
+  interface IBlockHeaderProps {
+    className?: string
+    title: string
+    tooltip?: React.ReactNode
+    content?: React.ReactNode,
+    position?: string,
+    prefix?: string
+  }
+
+  export class BlockHeader extends React.Component<IBlockHeaderProps, any> {}
+}
