+++ packages/zent/typings/libs/Cascader.d.ts
@@ -1,0 +1,18 @@
+/// <reference types="react" />
+
+declare module 'zent/lib/cascader' {
+  interface ICascaderProps {
+    value?: Array<any>
+    options?: Array<any>
+    title?: Array<any>
+    onChange?: Function
+    loadMore?: () => Promise<any>
+    changeOnSelect?: boolean
+    placeholder?: string
+    prefix?: string
+    className?: string
+    popClassName?: string
+  }
+
+  export class Cascader extends React.Component<ICascaderProps, any> {}
+}
