+++ packages/zent/typings/libs/Loading.d.ts
@@ -1,0 +1,18 @@
+/// <reference types="react" />
+
+declare module 'zent/lib/loading' {
+  interface ILoadingProps {
+    show?: boolean
+    static?: boolean
+    height?: number
+    zIndex?: number
+    className?: string
+    containerClass?: string
+    prefix?: string
+  }
+
+  export class Loading extends React.Component<ILoadingProps, any> {
+    static on(): void
+    static off(): void
+  }
+}
